# docs/api/openapi.yaml
...
components:
  schemas:
...
    FlashcardResultIn:
      properties:
        correct:
          title: Correct
          type: boolean
      required:
      - correct
      title: FlashcardResultIn
      type: object
    FlashcardResultOut:
      properties:
        flashcard_id:
          title: Flashcard Id
          type: string
        node_id:
          title: Node Id
          type: string
        node_mastery:
          title: Node Mastery
          type: string
      required:
      - flashcard_id
      - node_id
      - node_mastery
      title: FlashcardResultOut
      type: object
...
    GraphConfirmIn:
      description: 预览确认时的三种调整：删误抽 / 调分类 / 改内容。
      properties:
        delete_keys:
          items:
            type: string
          title: Delete Keys
          type: array
        edits:
          additionalProperties:
            $ref: '#/components/schemas/NodeEdit'
          title: Edits
          type: object
        reassign:
          additionalProperties:
            type: string
          title: Reassign
          type: object
      title: GraphConfirmIn
      type: object
    GraphConfirmOut:
      properties:
        category_count:
          title: Category Count
          type: integer
        edge_count:
          title: Edge Count
          type: integer
        edited_nodes:
          title: Edited Nodes
          type: integer
        node_count:
          title: Node Count
          type: integer
        ok:
          title: Ok
          type: boolean
        reason:
          title: Reason
          type: string
      required:
      - ok
      - reason
      - node_count
      - category_count
      - edge_count
      - edited_nodes
      title: GraphConfirmOut
      type: object
    GraphDraftOut:
      properties:
        draft:
          additionalProperties: true
          title: Draft
          type: object
        job_id:
          title: Job Id
          type: string
        status:
          title: Status
          type: string
      required:
      - job_id
      - status
      - draft
      title: GraphDraftOut
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    JobCreate:
      description: 新建加工任务。``recite_align`` 背诵对齐；``graph_extract`` 图谱抽取。
      properties:
        file_id:
          title: File Id
          type: string
        overwrite:
          default: false
          title: Overwrite
          type: boolean
        project_id:
          title: Project Id
          type: string
        type:
          default: recite_align
          enum:
          - recite_align
          - graph_extract
          title: Type
          type: string
      required:
      - project_id
      - file_id
      title: JobCreate
      type: object
    JobOut:
      properties:
        checkpoint_json:
          additionalProperties: true
          title: Checkpoint Json
          type: object
        cost_estimate_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Cost Estimate Json
        created_at:
          format: date-time
          title: Created At
          type: string
        done_units:
          title: Done Units
          type: integer
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        failed_units_json:
          items: {}
          title: Failed Units Json
          type: array
        heartbeat_at:
          format: date-time
          title: Heartbeat At
          type: string
        id:
          title: Id
          type: string
        progress:
          title: Progress
          type: integer
        project_id:
          title: Project Id
          type: string
        status:
          enum:
          - queued
          - running
          - success
          - failed
          - cancelled
          - interrupted
          title: Status
          type: string
        tokens_used:
          title: Tokens Used
          type: integer
        total_units:
          title: Total Units
          type: integer
        type:
          title: Type
          type: string
        units_json:
          items: {}
          title: Units Json
          type: array
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - id
      - project_id
      - type
      - status
      - progress
      - total_units
      - done_units
      - failed_units_json
      - checkpoint_json
      - tokens_used
      - units_json
      - created_at
      - updated_at
      - heartbeat_at
      title: JobOut
      type: object
...
    MasteryIn:
      properties:
        mastery:
          enum:
          - 'no'
          - mid
          - 'yes'
          title: Mastery
          type: string
      required:
      - mastery
      title: MasteryIn
      type: object
...
    NodeMasteryOut:
      properties:
        id:
          title: Id
          type: string
        mastery:
          title: Mastery
          type: string
        updated_at:
          title: Updated At
          type: string
      required:
      - id
      - mastery
      - updated_at
      title: NodeMasteryOut
      type: object
...
    ProjectCreate:
      properties:
        api_profile_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Profile Id
        name:
          maxLength: 120
          minLength: 1
          title: Name
          type: string
        type:
          enum:
          - graph
          - recite
          title: Type
          type: string
      required:
      - name
      - type
      title: ProjectCreate
      type: object
...
info:
  description: '学习工作台 Strayt 服务端。


    单用户免账号：所有 `/api/v1` 端点需 `Authorization: Bearer <访问令牌>`。

    唯一豁免是 `/health`（存活探针）。


    错误体统一为 `{ "error": { "code", "message", "detail?" } }`，`code` 是稳定契约，客户端据此分支。'
  title: Strayt 学习工作台
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/flashcards/{flashcard_id}/result:
    post:
      description: 答对/答错推进一档掌握度（no↔mid↔yes），两端钳位。
      operationId: post_flashcard_result_api_v1_flashcards__flashcard_id__result_post
      parameters:
      - in: path
        name: flashcard_id
        required: true
        schema:
          title: Flashcard Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlashcardResultIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlashcardResultOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 自测结果回写
      tags:
      - 闪卡自测
  /api/v1/jobs:
    get:
      operationId: list_jobs_api_v1_jobs_get
      parameters:
      - in: query
        name: project_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - pattern: ^(queued|running|success|failed|cancelled|interrupted)$
            type: string
          - type: 'null'
          title: Status
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/JobOut'
                title: Response List Jobs Api V1 Jobs Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 任务列表
      tags:
      - 加工任务
    post:
      description: '入队并立即在服务端后台开跑，客户端可以关窗。


        任务表里同时写 ``file_id``，断点续跑靠它找回要加工的文件。

        图谱任务在创建时就校验 LLM 配置：没有模型就没有知识点，与其建一个

        必然会失败的活，不如一开始就拦住并说明原因。'
      operationId: create_job_api_v1_jobs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 新建加工任务并开跑
      tags:
      - 加工任务
...
  /api/v1/jobs/{job_id}/graph-confirm:
    post:
      description: 把草稿（带手动调整）一次性原子落库，覆盖项目旧图谱。
      operationId: confirm_graph_api_v1_jobs__job_id__graph_confirm_post
      parameters:
      - in: path
        name: job_id
        required: true
        schema:
          title: Job Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphConfirmIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphConfirmOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 预览确认并入库
      tags:
      - 知识图谱
  /api/v1/jobs/{job_id}/graph-draft:
    get:
      description: 抽取任务完成后的预览草稿（尚未落正式表）。
      operationId: get_graph_draft_api_v1_jobs__job_id__graph_draft_get
      parameters:
      - in: path
        name: job_id
        required: true
        schema:
          title: Job Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDraftOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 阅读取预览草稿
      tags:
      - 知识图谱
...
  /api/v1/nodes/{node_id}/mastery:
    put:
      description: 直接写掌握度（在线路径）。离线走弱同步 ``node_mastery`` 实体。
      operationId: put_node_mastery_api_v1_nodes__node_id__mastery_put
      parameters:
      - in: path
        name: node_id
        required: true
        schema:
          title: Node Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MasteryIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeMasteryOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 设置节点掌握度
      tags:
      - 知识图谱
...
  /api/v1/projects:
    get:
      operationId: list_projects_api_v1_projects_get
      parameters:
      - description: 回收站视图：只列已删项目
        in: query
        name: include_deleted
        required: false
        schema:
          default: false
          description: 回收站视图：只列已删项目
          title: Include Deleted
          type: boolean
      - in: query
        name: type
        required: false
        schema:
          anyOf:
          - pattern: ^(graph|recite)$
            type: string
          - type: 'null'
          title: Type
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectOut'
                title: Response List Projects Api V1 Projects Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 项目列表
      tags:
      - 项目
    post:
      operationId: create_project_api_v1_projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 新建项目
      tags:
      - 项目
...
  /api/v1/projects/{project_id}/category-rules:
    get:
      operationId: list_category_rules_api_v1_projects__project_id__category_rules_get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoryRuleOut'
                title: Response List Category Rules Api V1 Projects  Project Id  Category Rules Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 一级分类规则列表
      tags:
      - 知识图谱
    put:
      description: 整表覆盖：先删该项目全部规则再按传入顺序重建，规则按 id 升序生效。
      operationId: put_category_rules_api_v1_projects__project_id__category_rules_put
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRuleListIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoryRuleOut'
                title: Response Put Category Rules Api V1 Projects  Project Id  Category Rules Put
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 全量保存一级分类规则
      tags:
      - 知识图谱
  /api/v1/projects/{project_id}/flashcards:
    get:
      description: 按项目取卡组，可只取某掌握度（自测优先复习弱项）。
      operationId: list_flashcards_api_v1_projects__project_id__flashcards_get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      - in: query
        name: mastery
        required: false
        schema:
          anyOf:
          - enum:
            - 'no'
            - mid
            - 'yes'
            type: string
          - type: 'null'
          title: Mastery
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FlashcardOut'
                title: Response List Flashcards Api V1 Projects  Project Id  Flashcards Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 闪卡卡组
      tags:
      - 知识图谱
  /api/v1/projects/{project_id}/flashcards/generate:
    post:
      description: '从已确认节点确定性重建整副卡组：question=名称，answer=概要/摘录。


        不调 LLM —— 概要/摘录在抽取时已限"仅基于给定材料"，这里复用它。

        同一事务先清旧卡再全量插入（与图谱擦除重建同语义）。'
      operationId: generate_flashcards_api_v1_projects__project_id__flashcards_generate_post
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlashcardGenOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 按当前图谱重建闪卡（幂等）
      tags:
      - 知识图谱
...
  /api/v1/projects/{project_id}/graph:
    get:
      description: 确认入库后的完整图谱：分类 → 节点（带引文）→ 关联。
      operationId: get_graph_api_v1_projects__project_id__graph_get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Graph Api V1 Projects  Project Id  Graph Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 图谱全量拉取
      tags:
      - 知识图谱
  /api/v1/projects/{project_id}/mastery-stats:
    get:
      description: 按掌握度三级与一级分类做聚合，供图谱视图着色与进度概览。
      operationId: get_mastery_stats_api_v1_projects__project_id__mastery_stats_get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasteryStatsOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 掌握度统计
      tags:
      - 知识图谱
...